refactor(UserAPI): move `UserAPI` into its own file
